refactor(coordination): import from submodules, empty __init__
Drop the re-exports and __all__ from superset/coordination/__init__.py so
the package init carries only the docstring, per review feedback that we
don't use the __init__ re-export pattern here. All consumers now import
directly from the submodule that owns the symbol:

  - CoordinationService              -> superset.coordination.base
  - SignalListener                  -> superset.coordination.types
  - CoordinationBackendUnavailableError -> superset.coordination.exceptions

Test patch targets and docstring cross-refs updated to the .base. paths.
