+++ tests/integration_tests/tasks/test_sync_join_wait.py
@@ -91,7 +91,8 @@ def test_wait_for_completion_timeout(app_context, login_as, get_user) -> None:
     try:
         # Force polling mode by mocking distributed_coordination as None
         with patch(
-            "superset.coordination.CoordinationService.get_backend", return_value=None
+            "superset.coordination.base.CoordinationService.get_backend",
+            return_value=None,
         ):
             with pytest.raises(TimeoutError):
                 TaskManager.wait_for_completion(
